Добавление Сущности, репозитория, зависимостей; редактированиме файла README

+++ userService/src/main/java/com/maxima/userService/repository/UserEntityRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserEntityRepository extends JpaRepository<UserEntity, Integer> {
+public interface UserEntityRepository extends JpaRepository<UserEntity, Long> {
 
 }
